+++ src/actions/resetPasswordAction.ts
@@ -1,8 +1,8 @@
 "use server"
 
-import { resetPasswordSchema } from "@/app/lib/ahth"
-import { ahthAction } from "./auth.actions"
+import { resetPasswordSchema } from "@/app/lib/auth"
+import { authAction } from "./auth.actions"
 
 export const resetPasswordAction = async (_: unknown, formData: FormData) => {
-  return ahthAction(_, formData, resetPasswordSchema)
+  return authAction(_, formData, resetPasswordSchema)
 }
