+++ src/app/(auth)/signup/signup-result/layout.tsx
@@ -1,0 +1,13 @@
+import { ReactNode } from "react"
+
+interface SignupResultProps {
+  children: ReactNode
+}
+
+export default function SignupResultLayout({ children }: SignupResultProps) {
+  return (
+    <>
+      {children}
+    </>
+  )
+}
